requirements.txt
absl-py==0.8.1
astor==0.8.0
boto==2.49.0
boto3==1.10.4
botocore==1.13.4
cachetools==3.1.1
chardet==3.0.4
click==7.0
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-auth==1.6.3
google-pasta==0.1.7
grpcio==1.24.3
h11==0.8.1
h5py==2.10.0
httptools==0.0.13
idna==2.8
jinja2==2.10.3
jmespath==0.9.4
keras-applications==1.0.8
keras-preprocessing==1.1.0
markdown==3.1.1
markupsafe==1.1.1
nltk==3.4.5
numpy==1.17.3
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-multipart==0.0.5
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
starlette==0.12.10
tensorboard==2.0.0
tensorflow==2.0.0a0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.6
uvicorn==0.9.1
uvloop==0.13.0
websockets==8.0.2
werkzeug==0.16.0
wrapt==1.11.2